refactor noris module
