+++ nest-city-account/src/noris/noris.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common'
 
 import { PrismaModule } from '../prisma/prisma.module'
-import { NorisConnectionService } from './noris-connection.service'
-import { NorisDeliveryMethodService } from './noris-delivery-method.service'
-import { NorisEdeskService } from './noris-edesk.service'
-import { NorisValidatorService } from './noris-validator.service'
+import { NorisConnectionService } from './services/noris-connection.service'
+import { NorisDeliveryMethodService } from './services/noris-delivery-method.service'
+import { NorisEdeskService } from './services/noris-edesk.service'
+import { NorisValidatorService } from './services/noris-validator.service'
 
 @Module({
   imports: [PrismaModule],
