+++ nest-city-account/src/tasks/subservices/tax-delivery-methods-tasks.subservice.ts
@@ -7,7 +7,7 @@ import { MailgunService } from '../../mailgun/mailgun.service'
 import {
   NorisDeliveryMethodService,
   UpdateNorisDeliveryMethodsData,
-} from '../../noris/noris-delivery-method.service'
+} from '../../noris/services/noris-delivery-method.service'
 import { DeliveryMethod } from '../../noris/types/noris.enums'
 import { PdfGeneratorService } from '../../pdf-generator/pdf-generator.service'
 import { ACTIVE_USER_FILTER, PrismaService } from '../../prisma/prisma.service'
